Ship logs to logit
I set this user provided service up manually in the staging space. It's
shipping to the (also manually created) "GOV.UK Staging Corona Forms" logit
stack.

We might need to do some bits in the application to get it to print logs
in a way that's easy to index, but we can cross that bridge a bit later.