Add demo pages
This adds demo pages that add the responses to the session,
but doesn't do anything with them. It provides the basic
navigation and structure for further question pages.

It assumes there will be an 'ammend answers' page, but doesn't
include one.

The page styles are to be confirmed, so I've done bare bones
stuff for this commit.