Use log-formatter to log to JSON
This should make the logs more useful for Splunk etc.

I've also removed the modernizr include, because it was 404ing and
giving me annoying 404 logs. As far as I can tell we don't actually use
the modernizr JS anywhere (I might be missing some CSS rule that uses
the classes it writes on body though).