Set session[:nhs_number] to string not hash
We are not actually storing the nhs number in a hash in the session.
Setting the default to an empty hash causes the view to display the
literal string "{}" in the textbox by default.

In addition remove the call to initialise the session in `show`.
I don't think we need it because the view is using `session.dig("nhs_number")`
so an absense of value will not throw an error.