Add query string to 'Change answers' link
It's useful to know how many users are changing their answers once they're
reviewing them on the check answers page. We  currently can't do this as we
don't have analytics on this service.

Adding a query string to the links on the check answers page means we can
determine how many users visited the question pages from the check answers
page without compromising a user's privacy.