Do not show nhs_number question on check answers
Do not show the nhs_number question if the skipped that question.

Users can get to a point where they are simultaneously saying the don't
know their NHS number but also providing one. This is because, the
check answers page was showing every possible question, even the one the
user may have skipped.

The value for the nhs_number is cleared every time they answer "No" to
the "Do you know your NHS number?" question so it's safe to use it
in the conditional here.