Refactor date_of_birth to return a Time object
We were lucky that the order of the keys in the hash matched the order we wanted to display them.
This change makes that ordering explicit.

The rest of the conditional had to be refactored slightly to take into account the date of birth no longer
needing to be "joined" together.