Container image to run feature tests in pipeline
Build a container that includes:

- ruby
- google chrome
- the application source code
- the application's dependencies

This is everything we need to run the feature tests against a remote URL
(previously we couldn't do this because we didn't have a container with
both ruby and chrome in).

We can get concourse to build this and upload it to the private Amazon
ECR (elastic container registry) that comes with the govuk-tools
concourse team (this is specific to GDS RE's concourse - see
https://reliability-engineering.cloudapps.digital/continuous-deployment.html#services)

Once we've got an image, we can use that to run the feature tests
against the staging URL.

We shouldn't run these tests against production yet, because they submit
the form, so they'd end up polluting live data. We could flag them so
they don't submit in staging though.

It's possible that building this image on every commit to master (and
blocking the pipeline on a clean build of the image) might turn out to
slow things down unacceptably. If that's the case we can look at caching
a base image or some other optimisation.