Use controller_name and controller_path instead of constants
The constant PAGE maps to controller_name and the template name maps
to controller_path. Also move the render call to the parent controller
definition of show and then override where necessary.

The constant NEXT_PAGE can be replaced by the use of named url helpers
which are faster to generate than passing an options has to url_for.