Use respond_to to turn 500 errors into 406 errors
If someone attempts to use a format at the end of the url it will
currently generate a 500 error because Rails will raise a
ActionView::MissingTemplate exception. By wrapping the render calls
in a respond_to block these will be ActionController::UnknownFormat
exceptions instead and can be ignored by Sentry because the response
status code is 406 Not Acceptable.