Set config.hosts for all environments
If a user sets an X-Forwarded header in a request, this will in some
cases override the Host header in a Rails app, resulting in URL
redirection to a different website. If an attacker manages to poison a
cache with such a header, users will be redirected to malicious sites.

In https://github.com/rails/rails/issues/29893, the resolution is to set
config.hosts.  Therefore setting this for all environments in this app.