Perform validations before setting session values
We are currently saving invalid responses to the session hash before
doing validations.  We need to validate first, but still replay the
invalid response back to the user.  This also removes all use of the
session hash out of the views.

As a side effect of this change we have resolved all string vs symbols
issues for the stored values, using symbols in all places now.