Hacky solution to maintain data integrity for medical conditions
We have updated the potential responses for the medical conditions
question.  However the downstream data sync process cannot be updated
until after the Bank Holiday weekend.

Therefore, we need to override either of the two "yes" options with the
text that was previously present in order to maintain data integrity
downstream.

Updating the value at this point allows us to replay the user's correct
answer on the "check your answers" page, and also if they change their
answer.

This code must be removed once the downstream data processing has been
updated.