Add opt-in Google Analytics tracking
This adds Google Analytics to the form, to help us understand
how people are using the service to offer help, so the development
team can make improvements.

User tracking will only be enabled if the users opts in.

This uses the cookie banner component, and copies the approach
(cookie settings page) from GOV.UK for this service.

If a user opts in to usage tracking, we'll load Google Analytics
so we can see how users move through the form in aggregate, in
order to improve the application.

Every page will have a cookie banner until the user accepts or
rejects cookies.

https://trello.com/c/ytRb9HIU/152