Format telephone numbers as UK national numbers
Removes any formatting added to telephone numbers by the user (e.g.
dashes, parentheses around the area code and spaces), removes the
international dialling code and then reformats as a UK national number
with the spaces in the correct place.

This will ensure that all phone numbers are in the same format for use
by downstream data consumers.