Bugfix: return valid error message for large numbers
If a user gives us a huge number e.g. 99999999999
we should return a 422 error, since no number that
large would be a valid date.

Currently we return a 500 error since Date.valid_date?
fails to convert it to an integer for some reason.
I'm not sure why since 99999999999.to_i.class == Integer.

I think it's because of a bug (behaviour) in the valid_date?
method that raises a RangeError. It should just return
false in thise case:
https://github.com/ruby/ruby/blob/1a4f33e84e0633e8f953ee3cb06b42dcc952444a/numeric.c#L2935