Add tests for 500 response from Ordnance Survey
A bit of jiggery pokery was needed to generate a 500 error response
using VCR.

The 500 response cassette in VCR was created by having the test look
like this initially:

```
it "returns 500 for a request with an invalid method" do
  VCR.use_cassette "address/500_response" do
    postcode_url = https://api.ordnancesurvey.co.uk/places/v1/addresses/postcode?dataset=LPI&key=123&postcode=SW1A2AA
    response = Faraday.post(postcode_url)

    expect(response.status).to eq(500)
  end
end
```

Notice the `post` call. Once the cassette was generated, this test
was updated, and the `method` defined in the cassette was manually
updated to `get`.

Co-authored-by: gclssvglx <graham.lewis@digital.cabinet-office.gov.uk>