Retry on Notifications::Client::AuthError
As part of the request to Notify, we send a current timestamp as part of
a JWT token (this is done by the notifications-ruby-client gem for us).
If the timestamp is not within 30 seconds, Notify returns an
unauthorised error, which raises Notifications::Client::AuthError in our
applications.

We have found a small number of cases where the request to Notify takes
longer than 30 seconds to open, therefore the timestamp is now out of
date and Notify rejects the request.  In these cases, we should retry
the request to ensure the email is actually sent to the user and we
don't log the error unless it keeps occurring.