Add form_response schema
Things to point out:

1. The "label" of the radio button/checkbox field is being stored as the
answer when then the form is submitted, so the enums in the schema
reflect this. That means that the schema will need to be updated every
time the content is, or the validation will fail.

2. Where possible the pattern matching is the same as that in the code.
The exception to this is the postcode, as the schema validation wouldn't
allow lowercase values even though `/i` was being used. Rather than
spending a lot of time trying to get this working right now, the pattern
has been simplified in favour of getting the schema validation out
faster. This problem also occurred on the business volunteering form.