Add schema validation helper
This uses a similar approach to [govuk_schemas](https://github.com/alphagov/govuk_schemas/blob/master/lib/govuk_schemas/rspec_matchers.rb)
to test that the data is valid.

`fully_validate` returns a list of validation errors which can be used to populate errors messages to the user
or Sentry entries if necessary.