Log error if schema validation fails
Until we're sure that the schema 100% correct, we don't want to stop
users from being able to submit the form. However we still need to know
what part of the validation failed, so the validation failures are being
sent to Sentry.

Creates a custom error so that they can be easily grouped in Sentry.