Allow "extra" value for "medical_conditions"
In https://github.com/alphagov/govuk-coronavirus-vulnerable-people-form/pull/245/commits/451f5228ccdad116ac02658cc58815239deb8d5f,
a hack was introduced to override the "yes" answers that the user sees with the old value so that the
downstream data team have time to update their code.

That hack is still in place, so we need to accept this value in the schema.