Get email address suggestion
Uses the mailcheck gem

mailcheck returns `false` if it thinks the email address ok, and a
hashed suggestion if it think it isn't:

So, `Mailcheck.new.suggest("me@example..com")` would return:

`{:address=>"me", :domain=>"example.com", :full=>"me@example.com"}`

This could be returned as part of `validate_email_address`, and used
to populate the message we send to the user, however, we'd need to make
sure that we don't stop the user from submitting their original email
address if they think it's correct.

We're going to use it to display a suggestion to the user on the new
check contact details page if Mailcheck returns any suggestions.