Add Truemail config
Most of the defaults are being used.

The exception is `default_validation_type`. There are 3 ways to validate
emails using Truemail:

- Regex: The regex that Truemail uses seems to catch more than the
built-in Rails one, so at a minimum we could just use this.

- MX: Checks that a mail exchange for the email domain exists.

- SMTP: Checks that an email can be sent to the email address.

At the moment the config is set to `:mx` because this seems like a good
middle ground.

The comments are there to make it easier to update the config later.