Display a suggestion
Display a suggestion on the check_contact_details page if we think the
user may have made an error.

We need to set `@form_responses` in the show method so that we can use
it to get the email suggestion. It seemed nicer to do this than to use
the session directly. I tried calling `super` before getting the
suggestion, but this renders the page before the suggestion is generated.

Also adds view tests so we can be sure that the hint is only being
displayed if there is a suggestion.

To get this to work `form_tag` needed to be updated in the view,
otherwise the tests fail with:

```
Failure/Error: <%= form_tag({},

ActionView::Template::Error:
 No route matches {:action=>"check_contact_details", :controller=>"coronavirus_form"}
```