Fixed bug with check answers page
We noticed that UPRN was showing up on the check answers page and felt that this is not useful to user so removed it from the view only.

The link to change postcode was incorrect in order to fix this we had to update the en.yml file to use the same key convention as other routes.

Removed redirect_to if check answers seen for the postcode and address lookup pages. This means that a user can no longer change their address without changing the address as stored in the session.  In short, if a user wants to change their address or postcode they have to go through the support_address page before being redirected to the check answers page.

Co-authored-by: Jon Hallam <jonathan.hallam@digital.cabinet-office.gov.uk>