Fix Schema Error.
On the check answers page, a validation check was being performed on the non-sanitised version of the session. This failed validation due to having the keys `session_id`, `_csrf_token`, etc. still in the session and reported an error, but it did not present an error to the user.

This change uses the sanitised version of the session, with those keys removed and amends the spec.