Fix bug with passing on UPRN
When the user moves part of the chosen address around (from the County
to the 2nd address line for example) the UPRN can get removed.  We want
to maintain the UPRN as much as possible.

This change ensures the UPRN is maintained against changes that do not
directly effect the postcode or other fundamental address elements. This
should ensure that we pass on more UPRN's.