Add docs about how to deploy breaking changes
Trello: https://trello.com/c/2nrSdb3X

Adds docs about how to deploy breaking changes so that we don't have ask
people to deploy things out of hours.

Other approaches considered:
Before settling on just adding guidance, pinning the user to a specific
version of the application or the JSON schema was considered.

However because we use continuous deployment to deploy the app, there
would be an added overhead of updating the application version number
for all changes except content changes. Also it doesn't look like it's
possible to add a version property to a JSON schema yet.