Guard against nil fields
Calling presence on nil will return nil. Currently, nils are entering
the stored data and are passed downstream causing data quality issues.

In order to mitigate against this we return an empty string instead
of nil at this point.

This change adds an appropriate guard and drys up the code a little.