Add stripPII function
Add function to redact the `reference_number` and `contact_gp` parameters in
the URL query string so that it can be used with Google Analytics cross domain
tracking.

This searches the query string (aka `window.location.search`) for the keys
`reference_number` and `contact_gp`, and redacts their values. This ensures
that no other information gets lost from the query string.

Copy of work done in alphagov/govuk-coronavirus-business-volunteer-form#389

Co-authored-by: Graham Lewis <graham.lewis@digital.cabinet-office.gov.uk>